Update Puppeteer tasks and add hurricane image download
This commit adds an option to run Puppeteer in headless mode for certain tasks in the server script. It also introduces a feature to download the latest hurricane cone image from the NHC website. The scheduling for the download of the thunderstorm forecast has been modified to run more frequently, every 15 minutes instead of twice a day. This will help to ensure more up-to-date weather information is available.